requirements.txt
geffnet==0.9.8
Pillow==7.2.0
numpy>=1.18.0
torch>=1.5.0
torchvision>=0.6.0
